generalized UserLocationContext to UserInfoContext
* login status is now global state
