+++ resources/js/pages/Home.tsx
@@ -3,18 +3,13 @@ import { MapContainer, TileLayer } from 'react-leaflet'
 import { StrictMode } from 'react'
 import "leaflet/dist/leaflet.css";
 import NewPost from '../components/NewPost';
-import { useState } from 'react';
-import { LatLng } from 'leaflet';
 import MouseCoordinates from '../components/MouseCoordinates';
 import Navbar from '../components/Navbar';
 import CustomControls from '../components/CustomControls';
 import PostContainer from '../components/PostContainer'
-import { UserLocationProvider } from '../components/UserLocationContext';
+import { UserInfoProvider } from '../components/UserInfoContext';
 
 function App() {
-
-  const [userLocation, setUserLocation] = useState<LatLng | null>(null)
-
   return (
     <StrictMode>
       <MapContainer 
@@ -24,11 +19,11 @@ function App() {
       zoom={3}
       className="w-screen h-screen my-0 mx-0 px-0 py-0 absolute z-10"
       maxBounds={[[85,-180],[-85, 180]]}>
-        <UserLocationProvider>
+        <UserInfoProvider>
           <PostContainer/>
           <TileLayer className="z-0" noWrap={true} minZoom={2} url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png" />
           <CustomControls yPosition="top" xPosition="left" className='ml-2 mt-2'>
-            <Navbar userLocation={userLocation as LatLng}/>
+            <Navbar />
           </CustomControls>
           <CustomControls yPosition="bottom" xPosition="right" className='mr-5 mb-5'>
             <NewPost/>
@@ -36,7 +31,7 @@ function App() {
           <div className='leaflet-bottom leaflet-left ml-1 mb-1 bg-white px-1 rounded'>
             <MouseCoordinates/>
           </div>
-        </UserLocationProvider>
+        </UserInfoProvider>
       </MapContainer>
     </StrictMode>
   )
